Use Raul::TimeStamp (LV2 compatible typed 32:32 fixed timestamp) everywhere. Fix initial size of Patchage messages window. Machina disabled for now (transitioning to generic timestamps).
git-svn-id: http://svn.drobilla.net/lad@1133 a436a847-0d15-0410-975c-d299462d15a1

+++ raul/src/TimeSlice.cpp
@@ -20,7 +20,7 @@
 namespace Raul {
 
 
-TimeSlice::TimeSlice(double tick_rate, double bpm)
+TimeSlice::TimeSlice(uint32_t tick_rate, double bpm)
 	: _tick_rate(tick_rate)
 	, _beat_rate(60.0/bpm)
 	, _start_ticks(0)
